requirements.txt
bcrypt==3.2.0
cffi==1.15.0
numpy==1.21.3
pycparser==2.20
python-dotenv==0.19.1
six==1.16.0
urllib3==1.26.7
